feat: add integration tests for artist songs and CifraClub search functionality
